==> 03_flux_sampling.ipynb <==
import pandas as pd
import cobra

from cobra.io import load_model

import numpy as np
import pickle

import os.path

## Basic settings

folder = "models"
folder_samples = "samples"

## Importing the models

model_control = cobra.io.read_sbml_model(os.path.join(f'{folder}','model_control.xml'))
model_kd = cobra.io.read_sbml_model(os.path.join(f'{folder}','model_kd.xml'))

Get the reactions from each of the models and store them as two sets.

reacts_control = {r.id for r in model_control.reactions}
reacts_kd = {r.id for r in model_kd.reactions}

len(reacts_control)

len(reacts_kd)

## Basic analyses

### Jaccard index

Jaccard index is a statistical measure to compare two populations / samples.

It can be expressed as a fraction between the number of the reactions in both samples (intersection) and the number of different reactions in both samples (union):
$$J=\frac{|S_1 \cap S_2|}{|S_1 \cup S_2|}$$

The size of the intersection is in the best case equal to the size of the union. This means that the index will take values from the interval $[0,1]$. Higher index reflects higher similarity between the samples.

J = len(reacts_control & reacts_kd)/len(reacts_control | reacts_kd)
J

## Summary

model_control.summary()

model_kd.summary()

## Metabolic flux sampling

We will employ the function called `sample`, which can be be accessed through the COBRApy package ([link](https://cobrapy.readthedocs.io/en/latest/sampling.html)).

from cobra.sampling import sample

We need to specify the model, number of samples nad sampling method. 

**CAUTION: this step is time consuming! You can skip the cells below in read the data from a file that was prepared in advance.**

s_control = sample(model_control, 1000, method="achr") 

**CAUTION: this step is time consuming! You can skip the cells below in read the data from a file that was prepared in advance.**

s_kd = sample(model_kd, 1000, method="achr")

s_control.to_csv(os.path.join(f'{folder_samples}','samples_control.csv'), index=False)

s_kd.to_csv(os.path.join(f'{folder_samples}','samples_kd.csv')', index=False)

**Reading the results from a file (in a case the steps above were skipped).**

s_control2 = pd.read_csv(os.path.join(f'{folder_samples}','samples_control.csv'))
s_kd = pd.read_csv(os.path.join(f'{folder_samples}','samples_kd.csv'))

s_control2

s_kd
